moving Cell to gameComponent because Board is independant but Game is not
