+++ src/main/scala/de/htwg/se/ludo/model/boardComponent/BoardInterface.scala
@@ -1,5 +1,7 @@
 package de.htwg.se.ludo.model.boardComponent
 
+import de.htwg.se.ludo.model.gameComponent.CellInterface
+
 trait BoardInterface {
   val spots: Vector[CellInterface]
   val size: Int
@@ -8,9 +10,4 @@ trait BoardInterface {
 
   def cell(spot: Int): CellInterface
   def replaceCell(spot: Int, cell: CellInterface): BoardInterface
-}
-
-trait CellInterface {
-  def isSet:Boolean
-  def getValue:String
 }
