+++ src/main/scala/de/htwg/se/ludo/model/boardComponent/boardBaseImpl/BasicBoardConstraints.scala
@@ -1,5 +1,7 @@
 package de.htwg.se.ludo.model.boardComponent.boardBaseImpl
 
+import de.htwg.se.ludo.model.gameComponent.{Cell, EmptyCell}
+
 object BasicBoardConstraints {
   val fields = 72
   val filling: Cell = EmptyCell
